Move benchmark fixture to conftest
