+++ tests/benchmark-models/test_petab_benchmark_jax.py
@@ -12,7 +12,6 @@ from amici.jax.petab import run_simulations, JAXProblem
 from amici.petab.petab_import import import_petab_problem
 from amici.petab.simulations import simulate_petab, LLH, SLLH
 
-import test_petab_benchmark as common
 
 benchmark_problem = common.benchmark_problem
 settings = common.settings
